requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
langchain
langchain-core
langgraph
langchain-community
langchain-google-genai
requests